backend/requirements.txt
Flask==2.2.5
Flask-CORS==4.0.0
Pillow==9.5.0
requests==2.28.2
urllib3==1.26.18
python-dotenv==0.21.1
